Refactor of Table and Catalog
New classes:
- Table: now an abstract base class declaring the full public table API
- LocalTable: the concrete local-execution table (former Table implementation); InsertableTable and View now subclass it.
- TableProxy / InsertableTableProxy: handles for hosted tables; all operations currently raise NotImplementedError.
- CatalogBase: abstract catalog API; Catalog now subclasses it.
- CatalogProxy: RPC-delegated catalog; stub for now.

Refactored:
- catalog.Path now carries the catalog location (org/db) parsed from pxt://<org>[:<db>]/... URIs, with .catalog_uri and .is_local;
subsumes and replaces the old PxtUri.
- Runtime: single _catalog replaced by _catalogs: dict[str, CatalogBase] keyed by catalog URI, plus get_catalog(path) that routes to the
local Catalog or a CatalogProxy.
- pxt.* globals now route every catalog operation through get_runtime().get_catalog(path); they no longer touch table internals
