+++ pixeltable/catalog/__init__.py
@@ -1,10 +1,13 @@
 # ruff: noqa: F401
 
 from .catalog import Catalog, retry_loop
+from .catalog_base import CatalogBase
+from .catalog_proxy import CatalogProxy
 from .column import Column
 from .dir import Dir
 from .globals import (
     ColumnVersionMd,
+    DirEntry,
     IfExistsParam,
     IfNotExistsParam,
     MediaValidation,
@@ -13,11 +16,14 @@ from .globals import (
     is_valid_identifier,
 )
 from .insertable_table import InsertableTable
+from .insertable_table_proxy import InsertableTableProxy
+from .local_table import LocalTable
 from .path import Path
 from .schema_object import SchemaObject
 from .table import Table
 from .table_metadata import ColumnMetadata, IndexMetadata, TableMetadata, VersionMetadata
 from .table_path import TableMdPath, TablePath, TableVersionPath
+from .table_proxy import TableProxy
 from .table_version import TableVersion
 from .table_version_handle import ColumnHandle, TableVersionHandle
 from .update_status import RowCountStats, UpdateStatus
