+++ pixeltable/catalog/catalog.py
@@ -26,14 +26,15 @@ from pixeltable.utils import fault_injection
 from pixeltable.utils.exception_handler import run_cleanup
 from pixeltable.utils.fault_injection import FaultLocation
 
+from .catalog_base import CatalogBase
 from .column import Column
 from .dir import Dir
-from .globals import IfExistsParam, IfNotExistsParam, MediaValidation, QColumnId
+from .globals import DirEntry, IfExistsParam, IfNotExistsParam, MediaValidation, QColumnId
 from .insertable_table import InsertableTable
+from .local_table import LocalTable
 from .path import ROOT_PATH, Path
 from .schema_object import SchemaObject
-from .table import Table
-from .table_path import TableVersionPath
+from .table_path import TablePath, TableVersionPath
 from .table_version import TableVersion, TableVersionKey, TableVersionMd
 from .table_version_handle import TableVersionHandle
 from .tbl_ops import DeleteTableMdOp, OpStatus, TableOp
@@ -165,7 +166,7 @@ class PendingTableOpsError(Exception):
         self.tbl_id = tbl_id
 
 
-class Catalog:
+class Catalog(CatalogBase):
     """The functional interface to getting access to catalog objects
 
     All interface functions must be called in the context of a transaction, started with Catalog.begin_xact() or
@@ -208,7 +209,7 @@ class Catalog:
     # - mutable version of a table: version == None (even though TableVersion.version is set correctly)
     # - snapshot versions: records the version of the snapshot
     _tbl_versions: dict[TableVersionKey, TableVersion]
-    _tbls: dict[tuple[UUID, int | None], Table]
+    _tbls: dict[tuple[UUID, int | None], LocalTable]
     _in_write_xact: bool  # True if we're in a write transaction
     _x_locked_tbl_ids: set[UUID]  # Ids of tables exclusively locked for write in the current transaction
     _modified_tvs: set[TableVersionHandle]  # TableVersion instances modified in the current transaction
@@ -1190,16 +1191,6 @@ class Catalog:
         rows = get_runtime().conn.execute(stmt).all()
         return {r.tbl_id: r.errors for r in rows}
 
-    @dataclasses.dataclass
-    class DirEntry:
-        dir: schema.Dir | None
-        dir_entries: dict[str, Catalog.DirEntry]
-        table: schema.Table | None
-
-        # Only populated for table entries when get_dir_contents() was called with with_error_counts=True;
-        # None otherwise (including for directory entries).
-        table_error_count: int | None = None
-
     @retry_loop(for_write=False)
     def get_dir_contents(
         self, dir_path: Path, recursive: bool = False, with_error_counts: bool = False
@@ -1213,23 +1204,23 @@ class Catalog:
     ) -> dict[str, DirEntry]:
         """Returns a dict mapping the entry names to DirEntry objects"""
         conn = get_runtime().conn
-        result: dict[str, Catalog.DirEntry] = {}
+        result: dict[str, DirEntry] = {}
 
         q = sql.select(schema.Dir).where(schema.Dir.parent_id == dir_id)
         rows = conn.execute(q).all()
         for row in rows:
             dir = schema.Dir(**row._mapping)
-            dir_contents: dict[str, Catalog.DirEntry] = {}
+            dir_contents: dict[str, DirEntry] = {}
             if recursive:
                 dir_contents = self._get_dir_contents(dir.id, recursive=True, error_counts=error_counts)
-            result[dir.md['name']] = self.DirEntry(dir=dir, dir_entries=dir_contents, table=None)
+            result[dir.md['name']] = DirEntry(dir=dir, dir_entries=dir_contents, table=None)
 
         q = sql.select(schema.Table).where(self._active_tbl_clause(dir_id=dir_id))
         rows = conn.execute(q).all()
         for row in rows:
             tbl = schema.Table(**row._mapping)
             err_count = error_counts.get(tbl.id, 0) if error_counts is not None else None
-            result[tbl.md['name']] = self.DirEntry(dir=None, dir_entries={}, table=tbl, table_error_count=err_count)
+            result[tbl.md['name']] = DirEntry(dir=None, dir_entries={}, table=tbl, table_error_count=err_count)
 
         return result
 
@@ -1252,7 +1243,7 @@ class Catalog:
             # If dest_obj is not None, it means `if_exists='ignore'` and the destination already exists.
             # If src_obj is None, it means `if_not_exists='ignore'` and the source doesn't exist.
             # If dest_obj is None and src_obj is not None, then we can proceed with the move.
-            if isinstance(src_obj, Table):
+            if isinstance(src_obj, LocalTable):
                 self._move_table(src_obj._id, new_path.name, dest_dir._id)
             elif isinstance(src_obj, Dir):
                 self._move_dir(src_obj._id, new_path.name, dest_dir._id)
@@ -1286,7 +1277,7 @@ class Catalog:
         - if both add and drop (= two directories are involved), lock the directories in a pre-determined order
           (in this case, by name) in order to prevent deadlocks between concurrent directory modifications
         """
-        assert drop_expected in (None, Table, Dir), drop_expected
+        assert drop_expected in (None, LocalTable, Dir), drop_expected
         assert (add_dir_path is None) == (add_name is None)
         assert (drop_dir_path is None) == (drop_name is None)
         dir_paths: set[Path] = set()
@@ -1329,7 +1320,7 @@ class Catalog:
             if drop_obj is None and raise_if_not_exists:
                 raise excs.NotFoundError(excs.ErrorCode.PATH_NOT_FOUND, f'Path {drop_path!r} does not exist.')
             if drop_obj is not None and drop_expected is not None and not isinstance(drop_obj, drop_expected):
-                expected_name = 'table' if drop_expected is Table else 'directory'
+                expected_name = 'table' if drop_expected is LocalTable else 'directory'
                 raise excs.RequestError(
                     excs.ErrorCode.INVALID_ARGUMENT,
                     f'{drop_path!r} needs to be a {expected_name} but is a {drop_obj._display_name()}',
@@ -1389,7 +1380,7 @@ class Catalog:
         - raise_if_not_exists is True and the path does not exist
         - expected is not None and the existing object has a different type
         """
-        assert expected in (None, Table, Dir), expected
+        assert expected in (None, LocalTable, Dir), expected
 
         if path.is_root:
             # the root dir
@@ -1419,7 +1410,7 @@ class Catalog:
                 excs.ErrorCode.PATH_ALREADY_EXISTS, f'Path {path!r} is an existing {obj._display_name()}.'
             )
         elif obj is not None and expected is not None and not isinstance(obj, expected):
-            expected_name = 'table' if expected is Table else 'directory'
+            expected_name = 'table' if expected is LocalTable else 'directory'
             raise excs.RequestError(
                 excs.ErrorCode.INVALID_ARGUMENT,
                 f'{path!r} needs to be a {expected_name} but is a {obj._display_name()}.',
@@ -1428,7 +1419,7 @@ class Catalog:
 
     def get_table_by_id(
         self, tbl_id: UUID, version: int | None = None, ignore_if_dropped: bool = False
-    ) -> Table | None:
+    ) -> LocalTable | None:
         """Must be executed inside a transaction. Might raise PendingTableOpsError."""
         if (tbl_id, version) not in self._tbls:
             if version is None:
@@ -1448,7 +1439,7 @@ class Catalog:
         media_validation: MediaValidation,
         create_default_idxs: bool,
         is_versioned: bool,
-    ) -> tuple[Table, bool]:
+    ) -> tuple[LocalTable, bool]:
         """
         Creates a new InsertableTable at the given path.
 
@@ -1463,7 +1454,7 @@ class Catalog:
 
             existing = self._handle_path_collision(path, InsertableTable, False, if_exists)
             if existing is not None:
-                assert isinstance(existing, Table)
+                assert isinstance(existing, LocalTable)
                 return existing._id, False
 
             dir = self._get_schema_object(path.parent, expected=Dir, raise_if_not_exists=True)
@@ -1489,7 +1480,7 @@ class Catalog:
         self._roll_forward()
 
         @retry_loop(read_tbl_ids=[tbl_id])
-        def _get_tbl() -> tuple[Table, bool]:
+        def _get_tbl() -> tuple[LocalTable, bool]:
             tbl = self.get_table_by_id(tbl_id)
             _logger.info(f'Created table {tbl._name()!r}, id={tbl._id}')
             Env.get().console_logger.info(f'Created table {tbl._name()!r}.')
@@ -1500,7 +1491,7 @@ class Catalog:
     def create_view(
         self,
         path: Path,
-        base: TableVersionPath,
+        base: TablePath,
         select_list: list[tuple[exprs.Expr, str | None]] | None,
         where: exprs.Expr | None,
         sample_clause: 'SampleClause' | None,
@@ -1512,26 +1503,29 @@ class Catalog:
         custom_metadata: Any,
         media_validation: MediaValidation,
         if_exists: IfExistsParam,
-    ) -> Table:
+    ) -> LocalTable:
+        assert isinstance(base, TableVersionPath)
+        base_path = base
+
         @retry_loop(for_write=True)
         def create_fn() -> UUID:
-            if not is_snapshot and base.is_mutable():
+            if not is_snapshot and base_path.is_mutable():
                 # this is a mutable view of a mutable base; X-lock the base and advance its view_sn before adding
                 # the view
-                base_id = base.tbl_id
+                base_id = base_path.tbl_id
                 assert len(self._acquire_write_lock(tbl_id=base_id)) == 1, base_id
                 self._x_locked_tbl_ids.add(base_id)
-                base_tv = self._get_tbl_version(TableVersionKey(base.tbl_id, None), validate_initialized=True)
+                base_tv = self._get_tbl_version(TableVersionKey(base_path.tbl_id, None), validate_initialized=True)
                 self.mark_modified_tv(base_tv.handle)
                 base_tv.tbl_md.view_sn += 1
                 result = get_runtime().conn.execute(
                     sql.update(schema.Table)
                     .values({schema.Table.md: dataclasses.asdict(base_tv.tbl_md, dict_factory=schema.md_dict_factory)})
-                    .where(schema.Table.id == base.tbl_id)
+                    .where(schema.Table.id == base_path.tbl_id)
                 )
                 assert result.rowcount == 1, result.rowcount
 
-            existing = self._handle_path_collision(path, View, is_snapshot, if_exists, base=base)
+            existing = self._handle_path_collision(path, View, is_snapshot, if_exists, base=base_path)
             if existing is not None:
                 assert isinstance(existing, View)
                 return existing._id
@@ -1540,7 +1534,7 @@ class Catalog:
             assert dir is not None
             md, ops = View._create(
                 path.name,
-                base=base,
+                base=base_path,
                 select_list=select_list,
                 additional_columns=additional_columns,
                 predicate=where,
@@ -1560,14 +1554,14 @@ class Catalog:
 
         self._roll_forward_ids.clear()
         view_id = create_fn()
-        if not is_snapshot and base.is_mutable():
+        if not is_snapshot and base_path.is_mutable():
             # invalidate base's TableVersion instance, so that it gets reloaded with the new mutable view
-            self._clear_tv_cache(base.tbl_version.key)
+            self._clear_tv_cache(base_path.tbl_version.key)
 
         self._roll_forward()
 
         @retry_loop(read_tbl_ids=[view_id])
-        def _get_tbl() -> Table:
+        def _get_tbl() -> LocalTable:
             return self.get_table_by_id(view_id)
 
         return _get_tbl()
@@ -1601,15 +1595,15 @@ class Catalog:
             del self._tbl_versions[key]
 
     @retry_loop(for_write=False)
-    def get_table(self, path: Path, if_not_exists: IfNotExistsParam) -> Table | None:
+    def get_table(self, path: Path, if_not_exists: IfNotExistsParam) -> LocalTable | None:
         obj = self._get_schema_object(
-            path, expected=Table, raise_if_not_exists=(if_not_exists == IfNotExistsParam.ERROR)
+            path, expected=LocalTable, raise_if_not_exists=(if_not_exists == IfNotExistsParam.ERROR)
         )
         if obj is None:
             _logger.info(f'Skipped table {path!r} (does not exist).')
             return None
 
-        assert isinstance(obj, Table)
+        assert isinstance(obj, LocalTable)
         # We need to clear cached metadata from tbl_version_path, in case the schema has been changed
         # by another process.
         obj._tbl_version_path.clear_cached_md()
@@ -1620,7 +1614,7 @@ class Catalog:
         def drop_fn() -> None:
             tbl = self._get_schema_object(
                 path,
-                expected=Table,
+                expected=LocalTable,
                 raise_if_not_exists=(if_not_exists == IfNotExistsParam.ERROR and not force),
                 lock_parent=True,
                 lock_obj=False,
@@ -1628,14 +1622,14 @@ class Catalog:
             if tbl is None:
                 _logger.info(f'Skipped table {path!r} (does not exist).')
                 return
-            assert isinstance(tbl, Table)
+            assert isinstance(tbl, LocalTable)
             self._drop_tbl(tbl, force=force, is_replace=False)
 
         self._roll_forward_ids.clear()
         drop_fn()
         self._roll_forward()
 
-    def _drop_tbl(self, tbl: Table | TableVersionPath, force: bool, is_replace: bool) -> None:
+    def _drop_tbl(self, tbl: LocalTable | TableVersionPath, force: bool, is_replace: bool) -> None:
         """
         Drop the table (and recursively its views, if force == True).
 
@@ -2007,7 +2001,7 @@ class Catalog:
             row = conn.execute(q).one_or_none()
             return schema.Dir(**row._mapping) if row is not None else None
 
-    def _load_tbl(self, tbl_id: UUID, ignore_pending_drop: bool = False) -> Table | None:
+    def _load_tbl(self, tbl_id: UUID, ignore_pending_drop: bool = False) -> LocalTable | None:
         """Loads metadata for the table with the given id and caches it."""
         from .insertable_table import InsertableTable
         from .view import View
@@ -2084,7 +2078,7 @@ class Catalog:
         self._tbls[tbl_id, None] = view
         return view
 
-    def _load_tbl_at_version(self, tbl_id: UUID, version: int) -> Table | None:
+    def _load_tbl_at_version(self, tbl_id: UUID, version: int) -> LocalTable | None:
         from .view import View
 
         # Load the specified TableMd and TableVersionMd records from the db.
@@ -2457,7 +2451,7 @@ class Catalog:
         status = conn.execute(sql.delete(schema.Table).where(schema.Table.id == tbl_id))
         assert status.rowcount == 1, status.rowcount
 
-    def load_md_for_export(self, tbl: Table) -> list[TableVersionMd]:
+    def load_md_for_export(self, tbl: LocalTable) -> list[TableVersionMd]:
         """
         Load metadata for the given table along with all its ancestors. The values of TableMd.current_version and
         TableMd.current_schema_version will be adjusted to ensure that the metadata represent a valid (internally
@@ -2678,7 +2672,7 @@ class Catalog:
 
         # Check for circularity
         if base is not None:
-            assert isinstance(obj, Table)  # or else it would have been caught above
+            assert isinstance(obj, LocalTable)  # or else it would have been caught above
             if obj._id in tuple(version.id for version in base.get_tbl_versions()):
                 raise excs.RequestError(
                     excs.ErrorCode.UNSUPPORTED_OPERATION,
@@ -2696,7 +2690,7 @@ class Catalog:
                 )
             self._drop_dir(obj._id, path, force=True)
         else:
-            assert isinstance(obj, Table)
+            assert isinstance(obj, LocalTable)
             self._drop_tbl(obj, force=if_exists == IfExistsParam.REPLACE_FORCE, is_replace=True)
         return None
