+++ pixeltable/catalog/schema_object.py
@@ -36,15 +36,18 @@ class SchemaObject(abc.ABC):
                 return None
             return get_runtime().catalog.get_dir(dir_id)
 
-    def _path(self) -> str:
-        """Returns the path to this schema object. Raises TABLE_NOT_FOUND if dropped."""
-        dir_id = self._dir_id()
-        if dir_id is None:
-            # an instance that's in the process of getting dropped has dir_id unset
-            raise excs.table_was_dropped(self._id)
+    def _path(self) -> 'catalog.Path':
+        """Returns the path to this schema object. Raises TABLE_NOT_FOUND if dropped.
+
+        Resolves the whole path in a single read transaction so the result is a consistent snapshot.
+        """
         with get_runtime().catalog.begin_xact(for_write=False):
+            dir_id = self._dir_id()
+            if dir_id is None:
+                # an instance that's in the process of getting dropped has dir_id unset
+                raise excs.table_was_dropped(self._id)
             path = get_runtime().catalog.get_dir_path(dir_id)
-            return str(path.append(self._name()))
+            return path.append(self._name())
 
     @abc.abstractmethod
     def _display_name(self) -> str:
