+++ pixeltable/serving/deploy.py
@@ -133,7 +133,7 @@ def _tables_from_fastapi_app(env_cfg: config.DeploymentConfig, module_attr: str)
                     'deployment.',
                 )
             assert route.endpoint.tbl is not None  # It's always non-None for 'compute' routes
-            table_paths.add(route.endpoint.tbl._path())
+            table_paths.add(str(route.endpoint.tbl._path()))
 
     _logger.info(
         f'Validated service {module_attr!r} with {len(table_paths)} table(s) referenced by {len(app.routes)} route(s).'
@@ -168,13 +168,14 @@ def _collect_project_files(project_dir: Path, include: list[str] | None, exclude
 
 
 def _export_tables_md(table_paths: set[str]) -> dict[str, Any]:
-    # Get all tables mentioned by any route contained in this deployment.
+    # Get all tables mentioned by any route contained in this deployment. These are local tables.
     tables = [pxt.get_table(path) for path in sorted(table_paths)]
+    assert all(isinstance(tbl, pxt.catalog.LocalTable) for tbl in tables)
 
     # Get the md for all ancestors of all such tables.
     catalog = get_runtime().catalog
     with catalog.begin_xact(for_write=False):
-        tables_md = [catalog.load_md_for_export(tbl) for tbl in tables]
+        tables_md = [catalog.load_md_for_export(tbl) for tbl in tables if isinstance(tbl, pxt.catalog.LocalTable)]
 
     # The ancestor md is returned as: primary table first, followed by ancestors in descending order.
     # Reverse so that ancestors come first, then flatten and de-duplicate (since some tables might have common
